Re-wrote the renderer to support multiple objects which are specified in world space coordinates.
This could be considered as a milestone since the world get computed using the perspective and view matrices of the camera and no origin for individual objects have to be specified.

Also, introduced the X.buffer object to allow the renderer to handle multiple objects. All objects get saved in a hash map of the renderer using a unique Id. The required glBuffers for the vertices and the colors are saved in separate hash maps using the same unique Id but they are stored as X.buffer objects with static count and size attributes for quick access. All in all this should be a pretty fast and flexible way of rendering multiple objects in 2d and 3d.