Introduce X.shaders (shader pair) class which bundles a vertex and a fragment shader.
By default, simple shaders get configured. It is possible to configure custom and complex shaders by overloading the class. The shader pair has a simple validate() method which checks if the configured attributes and uniform locators match the shader sources.

The validate() method will be called by an X.renderer before a shader is used.