Improved the X.testing framework by adding selenium support.
This has the following advantage:

 python xtk-utils/build.py -t

can be executed to run all XTK tests in Google Chrome as well as in Firefox (if they are installed). The testing results are printed to the shell.

In terms of the integration a special chromedriver had to be included which is different for mac, win and linux. Anyway, all platforms should be supported now.