Change the interactor to add events to the canvas of the X.renderer instead of the container.
The canvas is equivalent to the actual viewport - the container is the div container in the html page. The idea behind is that it should be possible to add multiple X.renderers to one div in html (f.e. in a lightbox display mode).

Added X.renderer.canvas() for easy access to the canvas.