Finalized the X.camera.ZoomEvent mechanism.
This event gets triggers a zoom operation of the camera. Currently, it gets fired by the interactor when using the mousewheel on a X.renderer (fast zooming) or by holding the right mouse button and moving the mouse in Y-direction (fine zooming).

The implementation can be considered as 'best-practice' in event handling and should replace/template all former and future event handling.