Slim down event management mechanism.
By optimizing the way events are used (hopefully final :)), the bubbling target is disabled. This makes it possible to minimize the interface/constructors for the X.camera and the X.interactor to require only what they really need as parameters instead of passing the full-blown X.renderer. This also has advantages in terms of generalization. F.e. one could set up a renderer with a camera without an interactor and just control the camera via code.

Also, it is now possible to listen to X.interactor events in the sampleApp to perform custom actions. For example, listen to mouse-movement and mouse-clicks and trigger output etc. This was just hacked, so has to be stabilized :)