Add the render type property to X.object.
It is now possible to specify the render mode of an X.object by passing a value from X.object.type to the constructor. By default or in case of an invalid type, TRIANGLES render mode gets used.

This commit also includes some experiments with transparency/opacity of objects by modifying the X.shaders object. This is hacked.