Obtain a rendering order of objects based on their opacity.
To be able to support transparency (blending) properly, the most opaque objects have to be rendered first, the least opaque objects have to be rendered last. This is achieved by storing all associated objects of a X.renderer using an AVL tree. The comparator of the AVL tree is set to X.object.OPACITY_COMPARATOR which compares opacity values of two X.objects.

This commit also replaces the way object IDs are handled. Instead of letting the X.renderer manage a uniqueId map of all associated X.objects, now all X.base-derived objects in XTK have a unique ID (X.base.id()) which gets set in the constructor. This mechanism might be useful for other stuff as well. By doing so in this case, the rendering mechanism is more compact and should still be pretty fast since the X.object's ID is still used as a key in all hashmaps.

Background information in OpenGL FAQ:
http://www.opengl.org/resources/faq/technical/transparency.htm

Just while commiting, I noticed that the implemented mechansism is not perfect! Quote:
"Fully opaque primitives need to be rendered first, followed by partially opaque primitives in back-to-front order."

The back-to-front order is not realized yet!