Introduce X.cuboid and X.cube classes for easy creation of 3D shapes.
Both classes derive from X.object where X.cuboid is the superclass of X.cube (since the cube has the same edge lengths in all three directions).

To create one of the two, it is only required to specify a center point in 3D and the edge lengths (3 for a cuboid, 1 for a cube). Everything else gets setup automatically.

It is possible to configure a global color, different colors per side (new feature) or different colors for each vertex.

The setSpacing method does not behave correctly right now. It should decimate the number of vertices used but right now does not connect the decimated ones.. (looks good so :)).

Also, line rendering is not supported but should be easy to add. Right now, if activated, line rendering looks good but does not behave as expected..

Added a couple of cubes/cuboids to sampleApp to directly show the new features..

These experiments show the following:

1) the depth rendering (sorting etc.) does not work correctly yet. Whatever gets rendered last is on top of other stuff for the camera. We probably need to order _all_ vertices by depth before rendering based on the current camera position.

2) the loading seems a little slow.. there is probably a lot of room for optimization.

3) camera movements is def. not perfect yet.. we knew that already so..

Style check runs through on the new classes - tried to keep everything clean.