Enhance performance by reducing default number of vertices in X.cuboid.
Now there are three spacing values (X,Y,Z). They define how large the vertices are in each direction for rendering the cuboid/box. By default, there are set to 2*radiusX (or Y or Z) to use the minimal required number of vertices for rendering.

For example a box with radius 5,5,5 uses on each side 2 vertices with the edge length 10.

It is still possible to decimate the surface by choosing a lower spacing after the cuboid was created.