First working animation of a X.box with an enabled texture.
To achieve this, the setupObject_ method has to be split into it's components. This is better practice anyway, because then we can update the glBuffers on a one-by-one base instead of re-creating all glBuffers again when the object has changed. F.e. if the transform changes, only the transform-glBuffer gets re-created.