Change the transform from vertex attributes to vertex uniforms.
This simplifies the way the transform matrix for each object is passed to the shaders (a lot :D). Also, got rid of about 300 lines of code in X.renderer and X.shader and at the same reduce the loading time for a 20 MB .STL file by 2 seconds. That's how it should be :)