Major update to X.renderer, X.base and X.object to clean things up.
1. All X.base classes now have a property X.base._dirty which marks a class and its direct properties as old/modified/dirty. This property can only be set internally. All X.object classes have additionally a method X.object.modified() which fires a modified event.

Now, the X.renderer listens to all added object's modified events and checks if certain properties are 'dirty' to re-calculate or re-load them. In particular, this is now active for textures and external files since everything else is very fast to re-calculate but it can be extended in the future.

The X.loader now does not fire its own modified event anymore but just calls object.modified() when the loading is done.