Simplify the X.shaders class by using enums and an automatic validation method.
This makes it easier to add new shader attributes and uniforms and also simplifies the access from the X.renderer.

Also, the normal matrix is not need since it is the 3x3 version of the view matrix. This should gain us some FPS since before the inverse and transposed view matrix was calculated on each interaction in X.camera.