Cleanup, simplify and speed-up X.renderer.
By using a generalized approach to communicate with X.shaders, it is now only necessary to modify the X.renderer.render_() method when adding new Shader uniforms or attributes.

Additionally, all attribute pointers and uniform locations are now stored in a hashmap and are requested only once (when the shader is added). This should gain additional FPS.