Propagate the dirty/clean pattern among all core classes.
This means, if a direct attribute of an object is changed and this attribute is not an X.class, the this._dirty flag has to be set true. F.e. if the visibility of an X.object changes (visibility is a direct bool-attribute), then the dirty flag of this X.object is set to true. If the points of an X.object change and the points are a X.triplets container, then, the dirty flag of the X.triplets container is changed and not the one of the X.object.

Before, the file path was a direct attribute of X.object and this was not conform with texture, points, colors etc. so X.file was created which follows exactly the dirty pattern. Else wise, a new file would set the same dirty flag like toggling the visibility of an X.object and this would permit to distinguish between easy changes on the X.object and big changes which make f.e. loading necessary..

Also, remove all 'Fatal: ' strings from the classes and added it to the X.exception base class which should save some bytes :D.