Introduce new configuration pattern for X.renderer and X.interactor.
All configuration flags are inside an enum 'config'.

This means, it is now possible to do the following:

      var r = new X.renderer('r');
      r.config.PROGRESSBAR_ENABLED = false;
      r.init();
      r.interactor().config.MOUSEWHEEL_ENABLED = false;
      r.interactor().config.MOUSECLICKS_ENABLED = false;
      r.interactor().config.KEYBOARD_ENABLED = false;
      r.interactor().init();

This example disable the progress bar during loading and also disable the interactor for certain events.

All configuration options in XTK should follow this scheme.

Also, fixed a mis-placement bug in X.progressbar.