Add a copy constructor to X.object.
This goes along with changes to the underlying classes like X.triplets and X.transform to be able to copy data not only point to it.

It is now possible to create a copied version of an existing X.object by doing this:

  var copyObject = new X.object(existingObject);
  copyObject.transform().translateX(10);

The copyObject will look like the existingObject but is moved 10 in X direction.