Export the X.interactor.config.* enumeration.
Re-add the license to the easybuild.py script.

Rename the source xtk.js to X.js to avoid confusion with the build xtk.js.

Renamed the Build.xml to Build.xml.template and also the output CDash XML to XTKBuild.xml and added it to .gitignore.

Modified the pull-build-* scripts to not checkout Build.xml anymore since this is fixed with the commit above.