Introduce the X.volume.threshold(lower,upper) functionality.
After parsing a .NRRD file, the threshold is set to the scalar range of the volume.

The threshold can be either modified by calling X.volume.threshold(a,b) or by setting X.volume._lowerThreshold and X.volume._upperThreshold.

The actual thresholding is performed in the shaders.

Also, X.slice was extended to have a pointer to its parent volume. This means we do not need to save the threshold settings or scalar range for each slice but store it only once.