Enhance the xy2ijk conversion code and enable SliceNavigators.
It is now possible to hold 'shift' while moving over a 2D renderer to control the perpendicular slices as well. This is accompanied by "slice navigators" which are shown as a cross-hair.

Unfortunately, the ijk2xy conversion (the heart of this commit) still has problems when a sliceview was rotated or panned.