The recent LEAK FIX causes problems with SliceDrop.
From experience, listenOnce never really worked with Google Closure. If there is a leak - we need to remove the listeners manually.

The previous commit did not work when dragging files into SliceDrop (the 3D view just stayed blank).

@nicolasrannou
@rudolphpienaar

So now it works in SliceDrop again but we maybe still have a leak identified by Nicolas.